Cleaned up imports, replaced sdi_12_data_collector.py with modularized version
